refactor(common-utils): Split `Extensions.kt` and `Utils.kt`
Create various smaller files with more dedicated names for a better
overview. Apply some trivial alignments as part of that, like renaming
of variables or generalizing `Collection` to `Iterable`.
